Solution to Check Triangle Validity

+++ Check Triangle Validity@@6788dd63490637cfc44e1b93/code.c
@@ -1,8 +1,8 @@
 #include<stdio.h>
 int main()
 {
-  long int a,b,c;
- scanf("%ld%ld%ld",&a,&b,&c);
+  int a,b,c;
+ scanf("%d%d%d",&a,&b,&c);
  if ( a>0 && b>0 && c>0 && a + b + c == 180){
     printf("Valid\n");
  }
